backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database & ORM
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary>=2.9.10

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator>=2.0.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pyotp>=2.9.0
qrcode[pil]>=7.4.2


# AI & ML
openai>=1.104.2
langchain>=0.3.4
langchain-community>=0.3.3
langchain-openai>=0.3.35
langgraph>=0.2.0
faiss-cpu==1.9.0.post1
pypdf==5.0.1
tiktoken>=0.8.0

# Redis & Caching
redis>=5.0.1
aioredis>=2.0.1
fakeredis>=2.20.1

# Background Tasks
celery>=5.3.4
redis>=5.0.1

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# HTTP Client
httpx>=0.25.2

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Image Processing & Computer Vision (Hugging Face API)
pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0